Starting fresh with this PR

+++ scripts/remove/nextcloud.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 # nextcloud uninstaller
-echo_query "Please enter mysql root password so that nextcloud database and user can be dropped." "hidden"
-read -s 'password'
+host=$(mysql --execute="select host from mysql.user where user = 'nextcloud';" | grep -E "localhost|127.0.0.1")
+echo_info "Removing all Nextcloud data"
+mysql --execute="DROP DATABASE nextcloud;" && mysql --execute="DROP USER nextcloud@$host;"
+if [[ $? != "0" ]]; then 
+    echo_error "MySQL Drop failed. Please try again or investigate"
+    exit 1
+fi
+crontab -l -u www-data | grep -v nextcloud > /tmp/.lol
+crontab -u www-data /tmp/.lol
 rm -rf /srv/nextcloud
 rm /etc/nginx/apps/nextcloud.conf
 systemctl reload nginx
-host=$(mysql -u root --password="$password" --execute="select host from mysql.user where user = 'nextcloud';" | grep -E "localhost|127.0.0.1")
-mysql --user="root" --password="$password" --execute="DROP DATABASE nextcloud;"
-mysql --user="root" --password="$password" --execute="DROP USER nextcloud@$host;"
+
 rm /install/.nextcloud.lock
